Change the namespace of the Exporter classes
